requirements.txt
anthropic>=0.8.0
requests>=2.28.1
python-dotenv>=0.21.0
argparse>=1.4.0
playwright>=1.30.0 